requirements.txt
certifi==2021.5.30
numpy==1.21.1
Pillow==8.3.1
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
